feat: more testing
